# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/init/": {
      "get": {
        "tags": [
          "Utilities"
        ],
        "description": "Get init page",
        "responses": {
          "200": {
            "description": "Successfully retrieved init page"
          },
          "400": {
            "description": "Fail! Something went wrong"
          }
        }
      },
      "post": {
        "tags": [
          "Utilities"
        ],
        "description": "Populate data in the database",
        "responses": {
          "200": {
            "description": "Successfully populated database"
          },
          "400": {
            "description": "Fail! Something went wrong"
          }
        }
      }
    },
...
    "/categories/": {
      "get": {
        "tags": [
          "Categories"
        ],
        "description": "Get all categories",
        "responses": {
          "200": {
            "description": "Successfully retrieved all categories"
          },
          "400": {
            "description": "Fail! Something went wrong"
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "description": "Create a new category",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/createCategory"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully created a new category"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "put": {
        "tags": [
          "Categories"
        ],
        "description": "Edit an existing category",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/editCategory"
            }
          },
          {
            "name": "CategoryId",
            "in": "path",
            "description": "ID number of a specific category",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully edited the category"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/categories/{CategoryId}": {
      "delete": {
        "tags": [
          "Categories"
        ],
        "description": "Delete an existing category",
        "parameters": [
          {
            "name": "CategoryId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID number of a specific category"
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted the category"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/admin/categories": {
      "get": {
        "tags": [
          "Admin-Categories"
        ],
        "description": "Get all categories",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all categories"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Admin-Categories"
        ],
        "description": "Create a new category",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/createCategory"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully created a new category"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "createCategory": {
      "type": "object",
      "properties": {
        "Category": {
          "type": "string",
          "example": "Example Category"
        }
      },
      "required": [
        "Category"
      ]
    },
    "editCategory": {
      "type": "object",
      "properties": {
        "NewCategory": {
          "type": "string",
          "example": "Edited Category"
        }
      },
      "required": [
        "NewCategory"
      ]
    },
...
  }
}